Finished debugging 12

+++ 12/9.c
@@ -1,12 +1,15 @@
 #include <stdio.h>
+
 #define PI 3.14159
+
 int main() {
-	float r;
-    float c, s;
-	scanf("%g", &r);
+	double r;
+    double c, s;
+	scanf("%lf", &r);
     c = 2 * PI * r;
     s = PI * r *r;
-    printf("%.3f\n",c);
-    printf("%.3f\n",s);
-    return 0;
+    printf("%.3f\n", c);
+    printf("%.3f\n", s);
+    
+	return 0;
 }
